+++ dashboard/service.go
@@ -3,8 +3,11 @@ package dashboard
 import (
 	"context"
 	"fmt"
+	"sort"
 
 	"github.com/xescugc/rebost/config"
+	"github.com/xescugc/rebost/membership"
+	"github.com/xescugc/rebost/state"
 	"github.com/xescugc/rebost/storing"
 
 	kitlog "github.com/go-kit/kit/log"
@@ -13,7 +16,7 @@ import (
 // Service exposes the Dashboard service interface
 type Service interface {
 	// ListNodes returns the list of all the nodes configuration
-	ListNodes(context.Context) ([]*config.Config, error)
+	ListNodes(context.Context) ([]*Node, error)
 }
 
 type service struct {
@@ -21,6 +24,13 @@ type service struct {
 	cfg     *config.Config
 
 	logger kitlog.Logger
+}
+
+// Node defines the aggregation of other entities
+// to represent a Node we need here
+type Node struct {
+	Config config.Config
+	State  membership.State
 }
 
 // New returns an implementation of the Dashboard with
@@ -34,16 +44,44 @@ func New(cfg *config.Config, m storing.Membership, logger kitlog.Logger) Service
 	}
 }
 
-func (s *service) ListNodes(ctx context.Context) ([]*config.Config, error) {
+func (s *service) ListNodes(ctx context.Context) ([]*Node, error) {
 	// As the Nodes only have the other nodes not the current, we append the
 	// current node as a first thing
-	cfgs := []*config.Config{s.cfg}
+	nodes := []*Node{
+		&Node{
+			Config: *s.cfg,
+			State: membership.State{
+				Volumes: make(map[string]state.State),
+			},
+		},
+	}
+
+	for _, lv := range s.members.LocalVolumes() {
+		s, err := lv.GetState(ctx)
+		if err != nil {
+			return nil, err
+		}
+		nodes[0].State.Volumes[lv.ID()] = *s
+	}
+
 	for _, n := range s.members.Nodes() {
 		cfg, err := n.Config(ctx)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get config: %w", err)
 		}
-		cfgs = append(cfgs, cfg)
+		st, err := s.members.GetNodeState(cfg.Name)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get metadata: %w", err)
+		}
+		nodes = append(nodes, &Node{
+			Config: *cfg,
+			State:  *st,
+		})
 	}
-	return cfgs, nil
+
+	sort.Slice(nodes, func(i, j int) bool {
+		return nodes[i].Config.Name > nodes[j].Config.Name
+	})
+
+	return nodes, nil
 }