+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Client can be initialized with multiple hosts and will request one at a time in order as a load balancer
   [Issue#16](https://github.com/xescugc/rebost/issues/16)
+- Visualization of the size of the Nodes (used and total) and the size of the cluster (use and total) to the Dashboard
+  [Issue#49](https://github.com/xescugc/rebost/issues/49)
 
 ### Fixed